==> globalprep/spp/v2022/_setup/4a_set_up_ocean_area_and_mpa_pct.Rmd <==
---
title: 'Set up ocean area raster and MPA pct rasters'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../../workflow/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

# Summary

To define spatial areas more accurately, create rasters of WDPA and ocean area at Gall-Peters resolution of 0.25 km^2. These can then be aggregated back to Gall Peters 10 km^2 working rasters, summing the total ocean area and total protected area within each working cell.

* Ocean area - use Natural Earth 10 m land polygons; mask over cell ID raster to exclude land.

# Data Sources

### WDPA
# Data Source
**Reference**: IUCN and UNEP-WCMC (2022), The World Database on Protected Areas (WDPA) [On-line], June 2022. Cambridge, UK: UNEP-WCMC. Available at: www.protectedplanet.net.

**Downloaded**: June 6, 2022

**Description**:  Shapefile of World Database on Protected Areas

**Time range**: 1800 - 2021; some protected areas do not have an associated "status year" and are reported as year 0.

**Format**:  Shapefile

**File location**: `Mazu:git-annex/globalprep/_raw_data/wdpa_mpa/d2022/WDPA_Jun2022_Public_shp/`


# Setup

```{r}
knitr::opts_chunk$set(echo = TRUE, eval=FALSE)
```

``` {r setup, echo = FALSE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'Figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
library(sf)
library(fasterize)
library(dplyr)
library(rgdal)
library(here)
library(ggplot2)
library(stringr)
library(mapview)
library(terra)

source(here('workflow/R/common.R'))

goal     <- 'spp'
scenario <- 'v2022'
dir_goal <- here('globalprep', goal, scenario)

### goal specific folders and info
dir_data  <- file.path(dir_goal, '_data')
dir_setup <- file.path(dir_goal, '_setup')
dir_anx  <- file.path(dir_M, 'git-annex/globalprep')
dir_goal_anx <- file.path(dir_anx, goal, scenario, 'spp_risk_dists')
dir_raw_data_anx <- file.path(dir_anx, '_raw_data')

source(file.path(dir_setup, 'common_fxns.R'))

sf::sf_use_s2(FALSE)

```


# Methods

## Ocean area raster

* Project Natural Earth 10M ocean polygons into Gall-Peters equal area projection
* rasterize at 500 m resolution
* aggregate up to 10 km resolution


### Create a 10 km^2 Gall Peters projection cell ID raster

 - Note: Just copy this over from last years, it doesn't change. 
 
``` {r rasterize natural earth ocean polygons to Gall-Peters}

ocean_area_file <- file.path(dir_spatial, 'ocean_area_rast.tif')
cell_id_file    <- file.path(dir_spatial, 'cell_id_rast.tif')
# ocean_1km_file  <- file.path(dir_goal_anx, 'spatial', 'ocean_1km.tif')
ocean_500m_file  <- file.path(dir_goal_anx, 'spatial', 'ocean_500m.tif')
# 
# if(!file.exists(ocean_area_file)) {
#   
#   ocean_sf_wgs84 <- read_sf(file.path(dir_spatial, 'ne_10m_ocean', 'ne_10m_ocean.shp'))
#   
#   ### Annoyingly, this counts the Caspian Sea as ocean... so make a block,
#   ### cut it out, and go from there.
#   caspian_block <- matrix(c(45, 35, 45, 48, 56, 48, 56, 35, 45, 35), 
#                           ncol = 2, byrow = TRUE)
#   caspian_sf <- st_polygon(list(caspian_block)) %>% 
#     st_sfc(crs = 4326)  %>%
#     st_set_precision(1000000) %>%
#     st_make_valid()
#   
#   
# ocean_sf <- ocean_sf_wgs84 %>%
#     st_difference(caspian_sf) %>%
#     st_transform(gp_proj4)
# 
#   bbox_ocean <- st_bbox(ocean_sf)
#   ext_r  <- extent(bbox_ocean$xmin, bbox_ocean$xmax, bbox_ocean$ymin, bbox_ocean$ymax)
#   rast_base_500m <- raster(ext_r, crs = crs(gp_proj4), res = 500)
#   rast_base_10km <- raster(ext_r, crs = crs(gp_proj4), res = 10000)
#   
#   if(!file.exists(ocean_500m_file)) {
#     rast_ocean_500m <- fasterize(ocean_sf, rast_base_500m, 
#                               field = NULL,   ### all polygons given value of 1
#                               background = 0) ### all non-polygons given value of 0
#     writeRaster(rast_ocean_500m, 
#                 ocean_500m_file,
#                 progress = 'text')
#     #plot(rast_ocean_500m) #check if worked in R 
#   } else {
#     rast_ocean_500m <- raster(ocean_500m_file)
#   }
#   
#   rast_ocean_10km <- raster::aggregate(rast_ocean_500m, 
#                                      fact = 20,   ### factor of 20 aggregation
#                                      fun = mean,  ### mean of 1s (ocean) and 0s (land)
#                                      progress = 'text')
#   
#   rast_ocean_area <- rast_ocean_10km * 100 ### multiply by 100 km^2 per cell
#   values(rast_ocean_area)[values(rast_ocean_area) < 1e-5] <- NA
#   
#   writeRaster(rast_ocean_area, ocean_area_file, overwrite = TRUE)
# }
# 
# if(!file.exists(cell_id_file)) {
#   rast_cell_id <- raster(ocean_area_file) %>%
#     setValues(1:ncell(.))
#   
#   writeRaster(rast_cell_id, cell_id_file)
# }

```

## Create WDPA MPA percent protection data frame

Marine regions are any protected area with MARINE flag of 1 or 2, or non-zero GIS_M_AREA.  Classify by IUCN category (I-VI -> 1-6; non-classified get 8).  Include only designated, adopted, and established parks.  Omit non-MPA management plans from U.S.  Rasterize to CHI resolution, calculate protected cells vs total cells to get percent of protection for each IUCN protection category.

Because we are rasterizing the protected areas down to 500 m resolution, we risk dropping MPAs smaller than about 1 km^2.  We assume these are not likely to have a significant impact on biodiversity protection, but see the sensitivity analysis.

### Create WDPA map at GP 500 m (.25 km^2) resolution

Actually, create two maps: one by protection category, another by year for no-take only (categories Ia, Ib, II, and no-take).

``` {r rasterize_wdpa_to_GP_500m_x_500m}

rast_cell_id    <- rast(cell_id_file)
rast_ocean_500m  <- rast(ocean_500m_file)
#plot(rast_ocean_500m) #check to see if loaded properly 

wdpa_marine_file <- file.path(dir_raw_data_anx, 'wdpa_mpa/d2022/wdpa_mar_by_cat_Jun2022_500m.tif')
wdpa_marine_file_int <- file.path(dir_raw_data_anx, 'wdpa_mpa/d2022/wdpa_mar_by_cat_Jun2022_500m_int.tif')


wdpa_poly_file_0   <- file.path(dir_raw_data_anx, 'wdpa_mpa/d2022/WDPA_Jun2022_Public_shp/WDPA_Jun2022_Public_shp_0', 
                              'WDPA_Jun2022_Public_shp-polygons.shp')
wdpa_poly_file_1   <- file.path(dir_raw_data_anx, 'wdpa_mpa/d2022/WDPA_Jun2022_Public_shp/WDPA_Jun2022_Public_shp_1', 
                              'WDPA_Jun2022_Public_shp-polygons.shp')
wdpa_poly_file_2   <- file.path(dir_raw_data_anx, 'wdpa_mpa/d2022/WDPA_Jun2022_Public_shp/WDPA_Jun2022_Public_shp_2', 
                              'WDPA_Jun2022_Public_shp-polygons.shp')

if(!file.exists(wdpa_marine_file)) {
  
  wdpa_poly_0 <- sf::st_read(wdpa_poly_file_0)
  wdpa_poly_1 <- sf::st_read(wdpa_poly_file_1)
  wdpa_poly_2 <- sf::st_read(wdpa_poly_file_2)
  
  wdpa_poly <- rbind(wdpa_poly_0, wdpa_poly_1, wdpa_poly_2)

   #wdpa_poly$IUCN_CAT %>% unique() 
  ##only grab ones that are reported/assigned/applicable
  iucn_cats <- c('Ia'  = 1,
                 'Ib'  = 1,
                 'II'  = 2,
                 'III' = 3,
                 'IV'  = 4,
                 'V'   = 5,
                 'VI'  = 6)
   wdpa_poly$STATUS %>% table()
     # Adopted   Designated  Established    Inscribed Not Reported     Proposed 
     #      34       254937         2401          250           34         1219 
  # wdpa_poly$NO_TAKE %>% unique()
  # x <- wdpa_poly %>% filter(NO_TAKE %in% c('All', 'Part'))
  # x <- wdpa_poly %>% filter(WDPAID == 309888)
  
  wdpa_marine <- wdpa_poly %>%
    filter(MARINE > 0 | GIS_M_AREA > 0) %>%
    filter(STATUS %in% c('Designated', 'Adopted', 'Established')) %>%
      ### no paper parks!
    filter(!str_detect(tolower(MANG_PLAN), 'non-mpa')) %>%
      ### omit non-MPA fisheries or species management plans!
    mutate(NO_TAKE = ifelse(WDPAID == 309888, TRUE, NO_TAKE),
             ### patch PIPA to be no take - currently "not reported"
           no_take = (NO_TAKE == 'All') | (NO_TAKE == 'Part' & NO_TK_AREA > 0.75 * GIS_M_AREA),
             ### if NO_TK_AREA is 75% or more of GIS area, count it...
           cat = iucn_cats[IUCN_CAT],
           cat = ifelse(no_take & !cat %in% 1:2, -1, cat), ### use -1 as a "no take" flag
           cat = ifelse(is.na(cat), 8, cat)) %>%           ### use 8 as an "other protected" flag
    arrange(cat) %>%
    st_transform(crs = gp_proj4)

  
  wdpa_marine_vect <- vect(wdpa_marine)

  rm(wdpa_marine)
  rm(wdpa_poly_0)
  rm(wdpa_poly_1)
  rm(wdpa_poly_2)
  rm(wdpa_poly) ## free up some space
    
  ### rasterize all protected areas to 500m global raster;
  ### use min value: will prioritize no-take and higher levels of protection
  terra::rasterize(x = wdpa_marine_vect, 
                                        y = rast_ocean_500m,
                                        field  = 'cat', 
                                        fun    = 'min', 
                                        filename = wdpa_marine_file_int)   

  #plot(wdpa_all_rast_500m)
  
  wdpa_all_rast_500m <- rast(wdpa_marine_file_int)
  
  rm(wdpa_marine_vect) ## free up some space
  
  rast_ocean_500m <- terra::classify(rast_ocean_500m ,cbind(-Inf, 0.005, NA), right=FALSE)
  # plot(rast_ocean_500m)
  
  wdpa_marine_rast_500m <- terra::mask(wdpa_all_rast_500m, rast_ocean_500m)
  
  # plot(wdpa_marine_rast_500m)
  # plot(rast(file.path(dir_raw_data_anx, 'wdpa_mpa/d2021/wdpa_mar_by_cat_Feb2021_500m.tif')))
  
  writeRaster(wdpa_marine_rast_500m, wdpa_marine_file, 
              overwrite = TRUE)
  
}

```

``` {r rasterize_protection_yr_to_GP_500m}

# rast_cell_id    <- rast(cell_id_file)
rast_ocean_500m  <- rast(ocean_500m_file)
# plot(rast_ocean_500m)

wdpa_prot_yr_file <- file.path(dir_raw_data_anx, 'wdpa_mpa/d2022', 
                               'wdpa_mar_by_prot_yr_Jun2022_500m.tif')

wdpa_prot_yr_file_int <- file.path(dir_raw_data_anx, 'wdpa_mpa/d2022/wdpa_mar_by_prot_yr_Jun2022_500m_int.tif')

wdpa_poly_file_0   <- file.path(dir_raw_data_anx, 'wdpa_mpa/d2022/WDPA_Jun2022_Public_shp/WDPA_Jun2022_Public_shp_0', 
                              'WDPA_Jun2022_Public_shp-polygons.shp')
wdpa_poly_file_1   <- file.path(dir_raw_data_anx, 'wdpa_mpa/d2022/WDPA_Jun2022_Public_shp/WDPA_Jun2022_Public_shp_1', 
                              'WDPA_Jun2022_Public_shp-polygons.shp')
wdpa_poly_file_2   <- file.path(dir_raw_data_anx, 'wdpa_mpa/d2022/WDPA_Jun2022_Public_shp/WDPA_Jun2022_Public_shp_2', 
                              'WDPA_Jun2022_Public_shp-polygons.shp')

if(!file.exists(wdpa_prot_yr_file)) {
  
  wdpa_poly_0 <- sf::st_read(wdpa_poly_file_0)
  wdpa_poly_1 <- sf::st_read(wdpa_poly_file_1)
  wdpa_poly_2 <- sf::st_read(wdpa_poly_file_2)
  
  wdpa_poly <- rbind(wdpa_poly_0, wdpa_poly_1, wdpa_poly_2)

   #wdpa_poly$IUCN_CAT %>% unique()
  iucn_cats <- c('Ia'  = 1,
                 'Ib'  = 1,
                 'II'  = 2,
                 'III' = 3,
                 'IV'  = 4,
                 'V'   = 5,
                 'VI'  = 6)
    #wdpa_poly$STATUS %>% table()
         # Adopted   Designated  Established    Inscribed Not Reported     Proposed 
         #  34       242940         1148          249           34         1201 
  # wdpa_poly$NO_TAKE %>% unique()
  # x <- wdpa_poly %>% filter(NO_TAKE %in% c('All', 'Part'))
  
  wdpa_no_take <- wdpa_poly %>%
    filter(MARINE > 0 | GIS_M_AREA > 0) %>%
    filter(STATUS %in% c('Designated', 'Adopted', 'Established')) %>%
      ### no paper parks!
    filter(!str_detect(tolower(MANG_PLAN), 'non-mpa')) %>%
      ### omit non-MPA fisheries or species management plans!
    mutate(NO_TAKE = ifelse(WDPAID == 309888, TRUE, NO_TAKE),
             ### patch PIPA to be no take - currently "not reported"
           no_take = (NO_TAKE == 'All') | (NO_TAKE == 'Part' & NO_TK_AREA > 0.75 * GIS_M_AREA),
             ### if NO_TK_AREA is 75% or more of GIS area, count it...
           cat = iucn_cats[IUCN_CAT],
           cat = ifelse(no_take & !cat %in% 1:2, -1, cat), ### use -1 as a "no take" flag
           cat = ifelse(is.na(cat), 8, cat)) %>%           ### use 8 as an "other protected" flag
    arrange(cat) %>%
    filter(cat <= 2) %>% ### this limits to categories Ia, Ib, II, and no-take
    st_transform(crs = gp_proj4)
  
  rm(wdpa_poly_0 , wdpa_poly_1, wdpa_poly_2, wdpa_poly)
  
  wdpa_no_take_vect <- vect(wdpa_no_take)
  
  rm(wdpa_no_take, wdpa_marine_rast_500m)
  
  ### rasterize all protected areas to 500 m global raster;
  ### use min value: will prioritize no-take and higher levels of protection
           wdpa_all_rast_500m <-   terra::rasterize(x = wdpa_no_take_vect, 
                                        y = rast_ocean_500m,
                                        field  = 'STATUS_YR', 
                                        fun    = 'min')
           
       #    writeRaster(wdpa_all_rast_500m, wdpa_prot_yr_file_int)
  
  rm(wdpa_no_take_vect)
  
  rast_ocean_500m <- terra::classify(rast_ocean_500m ,cbind(-Inf, 0.005, NA), right=FALSE)
  
    ### in case of floating point not-quite-zeros...
  
  wdpa_all_rast_500m <- rast(wdpa_prot_yr_file_int)
  
   terra::mask(wdpa_all_rast_500m, rast_ocean_500m, filename = wdpa_prot_yr_file)
   
   
   # wdpa_yr_rast_500m <- rast(wdpa_prot_yr_file)
    #plot(wdpa_yr_rast_500m)
  #plot(raster(file.path(dir_raw_data_anx, 'wdpa_mpa/d2020/wdpa_mar_by_prot_yr_Apr2020_500m.tif')))
  
  # writeRaster(wdpa_yr_rast_500m, wdpa_prot_yr_file, 
  #             overwrite = TRUE)
  # 
}

```

### Crosstab WDPA map to cell IDs

Be cautious when using `raster::crosstab()` as it returns factors that need to be coerced into integers... This code results in a file of cell IDs, IUCN protected area categories (1 = Ia and Ib, 2-6 = II - VI, 7 = no take, -1 = other), and proportion of cell protected by that category.

We can't simply aggregate upwards here, since we need to keep track of different protection levels (for the first) and year (for the second).

``` {r crosstab mpa areas}

wdpa_marine_rast_500m <- rast(wdpa_marine_file)
rast_cell_id    <- rast(cell_id_file)    ### cell IDs at 10 km^2
rast_ocean_500m <- rast(ocean_500m_file)  ### ocean presence at 1 km^2
rast_ocean_area <- rast(ocean_area_file) ### ocean proportional area at 10 km^2

cell_id_500m_file <- file.path(dir_goal_anx, 'spatial/cell_id_500m.tif')

### Need a 500 m raster of cell IDs for crosstabbing...
if(!file.exists(cell_id_500m_file)) {
  rast_cell_id_500m <- terra::resample(x = rast_cell_id, y = rast_ocean_500m,
                             filename = cell_id_500m_file,
                             method = 'near',
                             overwrite = TRUE)
} else {
  rast_cell_id_500m <- rast(cell_id_500m_file)
}
#plot(rast_cell_id_500m)
#plot(raster(file.path("/home/shares/ohi/git-annex/globalprep/spp/v2020/spp_risk_dists/spatial/cell_id_500m.tif")))

wdpa_cell_area_file <- file.path(dir_spatial, 'wdpa_mpa_area.csv')

### now break into chunks and mclapply across them to crosstab WDPA categories to cell IDs
if(!file.exists(wdpa_cell_area_file)) {
  ### Break into smaller chunks for faster processing and eventual parallelizing
  n_chunks <- 60
  
  crosstab_chunk <- function(rast1, rast2, chunk, n_chunks) {
    ## chunk <- 2
    # rast1 = wdpa_marine_rast_500m
    # rast2 = rast_cell_id_500m
    
    rast1 <- raster(rast1)
    rast2 <- raster(rast2)
    
    chunk_size <- ceiling(ncol(rast1) / n_chunks)
    left_bound <- (chunk - 1) * chunk_size + 1
    right_bound <- min(left_bound + chunk_size - 1, ncol(rast1))
    
   # chunk_ext <- terra::ext(xFromCol(rast1, left_bound), xFromCol(rast1, right_bound), yFromRow(rast1, nrow(rast1)), yFromRow(rast1, 1))
      
  chunk_ext <- extent(rast1, 1, nrow(rast1), left_bound, right_bound)
  
    message('Processing ', chunk, ': ', paste0(as.character(round(chunk_ext)), collapse = ', '))
    wdpa_chunk    <- raster::crop(rast1, chunk_ext)
    cell_id_chunk <- raster::crop(rast2, chunk_ext)
    
    wdpa_cells <- raster::crosstab(wdpa_chunk, cell_id_chunk, 
                           long = TRUE) %>%
      setNames(c('wdpa_category', 'cell_id', 'n_prot')) %>%
      mutate(wdpa_category = as.integer(as.character(wdpa_category)),
             cell_id       = as.integer(as.character(cell_id)))
        ### wdpa_category and cell_id are crosstabbed as factors - first
        ### convert to character (to unfactorize it) then to integer.  Otherwise
        ### you end up with factor index, not actual cell ID or category.
    
    return(wdpa_cells)
  }

  ### Use the function in an mclapply call
  system.time({
    wdpa_cells_list <- parallel::mclapply(1:n_chunks, mc.cores = 12,
      FUN = function(x) crosstab_chunk(wdpa_marine_rast_500m, 
                                       rast_cell_id_500m, 
                                       chunk = x, n_chunks))
#     user   system  elapsed 
# 5335.773  588.892  550.586  - about ten minutes! 
    
  })
  
  
  wdpa_cells_df <- bind_rows(wdpa_cells_list)
  
  area_df <- data.frame(cell_id  = values(rast_cell_id),
                        area_km2 = values(rast_ocean_area)) %>%
    rename(cell_id = 1, area_km2 = 2)
  
  wdpa_area_df <- wdpa_cells_df %>%
    full_join(area_df, by = 'cell_id') %>%
    filter(!is.na(wdpa_category)) %>%
    mutate(prot_area_km2 = n_prot / 4, ### count of cells in 500 x 500 m raster, div by 4
           mpa_pct = prot_area_km2 / area_km2) %>%
    dplyr::select(wdpa_category, cell_id, mpa_pct, prot_area_km2) %>%
      ### mpa pct is relative to *ocean* area - not *total cell* area
    group_by(wdpa_category, cell_id) %>%
    summarize(mpa_pct = sum(mpa_pct), prot_area_km2 = sum(prot_area_km2)) %>%
    ungroup()
  
  # old <- read_csv(file.path(here(), "globalprep/spp/v2021/_spatial/wdpa_mpa_area.csv"))
  
  sum(old$prot_area_km2) # 26841457
  sum(wdpa_area_df$prot_area_km2) # 24690672 
  
  write_csv(wdpa_area_df, wdpa_cell_area_file)
  
} else {
  # git_prov(wdpa_cell_area_file, 'output')
}

```

``` {r crosstab mpa years}

wdpa_yr_rast_500m <- raster(wdpa_prot_yr_file)
rast_cell_id    <- raster(cell_id_file)    ### cell IDs at 10 km^2
rast_ocean_500m  <- raster(ocean_500m_file)  ### ocean presence at 1 km^2
rast_ocean_area <- raster(ocean_area_file) ### ocean proportional area at 10 km^2

cell_id_500m_file <- file.path(dir_goal_anx, 'spatial/cell_id_500m.tif')
rast_cell_id_500m <- raster(cell_id_500m_file)

wdpa_cell_year_file <- file.path(dir_spatial, 'wdpa_mpa_year.csv')

### now break into chunks and mclapply across them to crosstab WDPA categories to cell IDs
if(!file.exists(wdpa_cell_year_file)) {
  ### Break into smaller chunks for faster processing and eventual parallelizing
  n_chunks <- 60
  
  crosstab_chunk2 <- function(rast1, rast2, chunk, n_chunks) {
    ### chunk <- 30
    chunk_size <- ceiling(ncol(rast1) / n_chunks)
    left_bound <- (chunk - 1) * chunk_size + 1
    right_bound <- min(left_bound + chunk_size - 1, ncol(rast1))
    
    chunk_ext <- extent(rast1, 1, nrow(rast1), left_bound, right_bound)
    message('Processing ', chunk, ': ', paste0(as.character(round(chunk_ext)), collapse = ', '))
    wdpa_chunk    <- crop(rast1, chunk_ext)
    cell_id_chunk <- crop(rast2, chunk_ext)
    
    wdpa_cells <- crosstab(wdpa_chunk, cell_id_chunk, 
                           progress = 'text',
                           long = TRUE) %>%
      setNames(c('mpa_year', 'cell_id', 'n_prot')) %>%
      mutate(mpa_year = as.integer(as.character(mpa_year)),
             cell_id  = as.integer(as.character(cell_id)))
        ### mpa_year and cell_id are crosstabbed as factors - first
        ### convert to character (to unfactorize it) then to integer.  Otherwise
        ### you end up with factor index, not actual cell ID or category.
    
    return(wdpa_cells)
  }

  ### Use the function in an mclapply call
  system.time({
    wdpa_cells_list <- parallel::mclapply(1:n_chunks, mc.cores = 12,
      FUN = function(x) crosstab_chunk2(wdpa_yr_rast_500m, 
                                        rast_cell_id_500m, 
                                        chunk = x, n_chunks))
#     user   system  elapsed 
# 5401.628  619.072  562.306 - about 10 mins
  })
  
  
  wdpa_cells_df <- bind_rows(wdpa_cells_list)
  
  area_df <- data.frame(cell_id  = raster::values(rast_cell_id),
                        area_km2 = raster::values(rast_ocean_area))
  
  wdpa_year_df <- wdpa_cells_df %>%
    full_join(area_df, by = 'cell_id') %>%
    filter(!is.na(mpa_year)) %>%
    mutate(prot_area_km2 = n_prot / 4, 
              ### count of cells in 500 x 500 m^2 raster, times .25 km^2 area per cell
           mpa_pct = prot_area_km2 / area_km2) %>%
    dplyr::select(mpa_year, cell_id, mpa_pct, prot_area_km2) %>%
      ### mpa pct is relative to *ocean* area - not *total cell* area
    group_by(mpa_year, cell_id) %>%
    summarize(mpa_pct = sum(mpa_pct), prot_area_km2 = sum(prot_area_km2)) %>%
    ungroup()
  
  
  write_csv(wdpa_year_df, wdpa_cell_year_file)
  
} else {
  # git_prov(wdpa_cell_year_file, 'output')
}

```

### View WDPA protected area proportions as raster

Proportional protection can be multiplied by ocean area raster to get area of protection.  Collect by category (to avoid duplicate "by" values):

* No Take (categories 1 & 2 and specifically designated "no take")
    * classified as 1, 2, and -1 respectively
* Conservation (categories 1-4 plus no take)
    * classified as 1, 2, 3, 4, and -1 (no take)
* All protection (non-NA)
    * classified as 1-6, -1 (no take), and 8 (other protection)

Note: these are not saved as separate rasters at this point... they are easily created from the dataframe of cell ID to protected proportion.

``` {r}

wdpa_area_df <- read_csv(wdpa_cell_area_file)

# wdpa_area_df <- read_csv("/home/sgclawson/github/ohiprep_v2022/globalprep/spp/v2021/_spatial/wdpa_mpa_area.csv")

notake_df <- wdpa_area_df %>%
  filter(wdpa_category <= 2) %>%
  group_by(cell_id) %>%
  summarize(mpa_pct = sum(mpa_pct)) %>%
  rename(from = cell_id, to = mpa_pct)

conservation_df <- wdpa_area_df %>%
  filter(wdpa_category <= 4) %>%
  group_by(cell_id) %>%
  summarize(mpa_pct = sum(mpa_pct)) %>%
  rename(from = cell_id, to = mpa_pct)

all_prot_df <- wdpa_area_df %>%
  group_by(cell_id) %>%
  summarize(mpa_pct = sum(mpa_pct)) %>%
  rename(from = cell_id, to = mpa_pct)

notake_raster <- classify(x = rast(rast_cell_id), notake_df, othersNA = TRUE)
plot(notake_raster, main = 'No take protection (Ia, Ib, II + no take)')

conservation_raster <- classify(rast(rast_cell_id), conservation_df, othersNA = TRUE)
plot(conservation_raster, main = 'Conservation protection (Ia - IV + no take)')

all_prot_raster <- classify(rast(rast_cell_id), all_prot_df, othersNA = TRUE)
plot(all_prot_raster, main = 'All protection (any category)')
```

-----
